Resolved Review comment
