+++ Ginger/GingerAutoPilotTest/SwaggetTests.cs
@@ -111,7 +111,7 @@ namespace UnitTests.NonUITests.AutoPilot
 
             RequestToTest = requests.Where(x => x.Name == @"Add a new pet to the store-JSON").ElementAt(0);
             Assert.AreEqual(8, RequestToTest.ReturnValues.Count(), "SwaggerCheckResponseParameterCount");
-            Assert.AreEqual("/api/v3/pet", RequestToTest.EndpointURL);
+            Assert.AreEqual("/pet", RequestToTest.EndpointURL);
 
             RequestToTest = requests.Where(x => x.Name == @"Place an order for a pet-JSON").ElementAt(0);
             requestBody = @"{" + Environment.NewLine + "\"id\": <ID>," + Environment.NewLine + "\"petId\": <PETID>," + Environment.NewLine + "\"quantity\": <QUANTITY>," + Environment.NewLine + "\"shipDate\": \"<SHIPDATE>\"," + Environment.NewLine + "\"status\": \"<STATUS>\"," + Environment.NewLine + "\"complete\": <COMPLETE>" + Environment.NewLine + "}";
